+++ CHANGELOG.md
@@ -2,6 +2,7 @@ In Development
 --------------
 - `pyrepo release` will now error if the project uses versioningit and no
   explicit version or bump option is given on the command line
+- Use `tomllib` on Python 3.11
 
 v2022.10.16
 -----------